Move WorkSearchScheme into own file
